Migrate node schemas to new format
