+++ backend/templates/attributeschema_imports.j2
@@ -1,5 +1,6 @@
 from typing import Optional, Any, TYPE_CHECKING
 
+from infrahub.core.models import HashableModel
 from infrahub.core.schema.dropdown import DropdownChoice  # noqa: TCH001
 
 if TYPE_CHECKING:
