+++ backend/templates/basenodeschema_imports.j2
@@ -1,0 +1,7 @@
+from typing import Optional
+
+from infrahub.core.constants import BranchSupportType
+from infrahub.core.models import HashableModel
+from infrahub.core.schema.attribute_schema import AttributeSchema  # noqa: TCH001
+from infrahub.core.schema.filter import FilterSchema # noqa: TCH001
+from infrahub.core.schema.relationship_schema import RelationshipSchema  # noqa: TCH001
