+++ backend/templates/generate_schema.j2
@@ -1,13 +1,14 @@
+# Generated by "invoke backend.generate", do not edit directly
+
 from __future__ import annotations
 
 
 from pydantic import BaseModel, Field
-from infrahub.core.models import HashableModel
 
 {% include schema|lower + "_imports.j2" %}
 
 
-class Generated{{ schema }}(HashableModel):
+class Generated{{ schema }}({{ parent }}):
 {%- for attribute in node.attributes %}
     {{ attribute.name }}: {{ attribute.type_annotation }} = Field(
         {{ attribute.default_definition }},
