+++ backend/templates/nodeschema_imports.j2
@@ -1,0 +1,3 @@
+from typing import Optional
+
+from infrahub.core.schema.basenode_schema import BaseNodeSchema
