+++ backend/templates/relationshipschema_imports.j2
@@ -1,5 +1,5 @@
 from typing import Optional
 
 from infrahub.core.constants import BranchSupportType, RelationshipKind, RelationshipCardinality, RelationshipDirection # noqa: TCH001
-
+from infrahub.core.models import HashableModel
 from infrahub.core.schema.filter import FilterSchema # noqa: TCH001
