+++ pyproject.toml
@@ -251,6 +251,22 @@ module = "infrahub.core.schema"
 ignore_errors = true
 
 [[tool.mypy.overrides]]
+module = "infrahub.core.schema.attribute_schema"
+ignore_errors = true
+
+[[tool.mypy.overrides]]
+module = "infrahub.core.schema.basenode_schema"
+ignore_errors = true
+
+[[tool.mypy.overrides]]
+module = "infrahub.core.schema.node_schema"
+ignore_errors = true
+
+[[tool.mypy.overrides]]
+module = "infrahub.core.schema.relationship_schema"
+ignore_errors = true
+
+[[tool.mypy.overrides]]
 module = "infrahub.core.schema_manager"
 ignore_errors = true
